refactor: ステージ進捗の保存を分離
